+++ .php-cs-fixer.dist.php
@@ -16,7 +16,7 @@ $config
                 '='  => 'align',
             ],
         ],
-        'ordered_imports'         => true,
+        'ordered_imports'         => false,
         'global_namespace_import' => ['import_classes' => true, 'import_constants' => true, 'import_functions' => true],
     ])
     ->setFinder(