+++ src/Wildcard.php
@@ -33,7 +33,7 @@ declare(strict_types=1);
 
 /**
  * Esi\Wildcard is a fork of rkrx/php-wildcards (https://github.com/rkrx/php-wildcards) which is:
- *     Copyright (c) 2014-2023 Ronald Kirschler
+ *     Copyright (c) 2013-2023 Ronald Kirschler
  *
  * To see a list of changes in comparison to the original library {@see CHANGELOG.md}.
  */
@@ -43,10 +43,11 @@ namespace Esi\Wildcard;
 use Closure;
 use RuntimeException;
 
+use function call_user_func;
 use function explode;
 use function implode;
 use function ltrim;
-use function preg_last_error;
+use function preg_last_error_msg;
 use function preg_match;
 use function preg_quote;
 use function preg_replace;
@@ -58,12 +59,6 @@ use function str_ends_with;
 use function str_starts_with;
 use function strlen;
 
-use const PREG_BACKTRACK_LIMIT_ERROR;
-use const PREG_BAD_UTF8_ERROR;
-use const PREG_BAD_UTF8_OFFSET_ERROR;
-use const PREG_INTERNAL_ERROR;
-use const PREG_JIT_STACKLIMIT_ERROR;
-use const PREG_RECURSION_LIMIT_ERROR;
 use const PREG_SPLIT_DELIM_CAPTURE;
 
 /**
@@ -110,12 +105,13 @@ class Wildcard
         $pattern        = (string) preg_replace('/\\*+/', '*', $pattern);
         $doesNotContain = !str_contains($pattern, '?');
 
-        match (true) {
-            preg_match('/^[^\\*]+\\*$/', $pattern) === 1 && $doesNotContain        => $this->initStartsWith($pattern),
-            preg_match('/^\\*[^\\*]+$/', $pattern) === 1 && $doesNotContain        => $this->initEndsWith($pattern),
-            preg_match('/^[^\\*]+\\*[^\\*]+$/', $pattern) === 1 && $doesNotContain => $this->initStartsAndEndsWith($pattern),
-            default                                                                => $this->initRegExp($pattern)
+        $func = match (true) {
+            preg_match('/^[^\\*]+\\*$/', $pattern) === 1 && $doesNotContain        => 'initStartsWith',
+            preg_match('/^\\*[^\\*]+$/', $pattern) === 1 && $doesNotContain        => 'initEndsWith',
+            preg_match('/^[^\\*]+\\*[^\\*]+$/', $pattern) === 1 && $doesNotContain => 'initStartsAndEndsWith',
+            default                                                                => 'initRegExp'
         };
+        call_user_func([$this, $func], $pattern);
     }
 
     /**
@@ -166,17 +162,7 @@ class Wildcard
 
         if ($parts === false) {
             //@codeCoverageIgnoreStart
-            $pregError = match(preg_last_error()) {
-                PREG_INTERNAL_ERROR => 'internal PCRE error',
-                PREG_BACKTRACK_LIMIT_ERROR => 'backtrack limit was exhausted',
-                PREG_RECURSION_LIMIT_ERROR => 'recursion limit was exhausted',
-                PREG_BAD_UTF8_ERROR => 'malformed UTF-8 data ',
-                PREG_BAD_UTF8_OFFSET_ERROR => 'the offset didn\'t correspond to the begin of a valid UTF-8 code point',
-                PREG_JIT_STACKLIMIT_ERROR => 'failed due to limited JIT stack space',
-                default => 'unknown'
-            };
-
-            throw new RuntimeException(sprintf('PCRE error on preg_split: %s', $pregError));
+            throw new RuntimeException(sprintf('PCRE error on preg_split: %s', preg_last_error_msg()));
             //@codeCoverageIgnoreEnd
         }