+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 A not so exhaustive list of changes for each release.
 
 For a more detailed listing of changes between each version, 
-you can use the following url: https://github.com/ericsizemore/wildcard/compare/v1.0.0...v1.1.0. 
+you can use the following url: https://github.com/ericsizemore/wildcard/compare/v1.0.0...v1.0.1. 
 
 Simply replace the version numbers depending on which set of changes you wish to see.
+
+### 1.0.1 ()
+
+  * Updated unit tests to use `assertTrue` and `assertFalse` instead of `assertEquals`.
+  * Added the use of the PHPUnit `CoversClass` attribute.
+  * Minor CS fixes.
 
 ### 1.0.0 (2024-02-02)